orchestrator/pyproject.toml
[tool.poetry]
name = "orchestrator"
version = "0.1.0"
description = "The AI brain of the Sentinel system, handling workflows and LLM integration."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
# Web Framework
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
gunicorn = "^22.0.0"

# AI & LangChain
langchain = "^0.1.20"
langchain-openai = "^0.1.6"
langgraph = "^0.0.48"

# gRPC
grpcio = "^1.64.0"
grpcio-tools = "^1.64.0"
protobuf = "^4.25.3"

# Databases & Caching
chromadb = "^0.5.0"
sentence-transformers = "^2.7.0" # For ChromaDB embeddings
redis = {extras = ["hiredis"], version = "^5.0.4"}

# Utilities
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"